+++ pytorch_lightning/metrics/functional/self_supervised.py
@@ -15,10 +15,10 @@ import torch
 
 
 def embedding_similarity(
-        batch: torch.Tensor,
-        similarity: str = 'cosine',
-        reduction: str = 'none',
-        zero_diagonal: bool = True
+    batch: torch.Tensor,
+    similarity: str = 'cosine',
+    reduction: str = 'none',
+    zero_diagonal: bool = True
 ) -> torch.Tensor:
     """
     Computes representation similarity