+++ pytorch_lightning/metrics/utils.py
@@ -44,7 +44,11 @@ def _check_same_shape(pred: torch.Tensor, target: torch.Tensor):
 
 
 def _input_format_classification_one_hot(
-    num_classes: int, preds: torch.Tensor, target: torch.Tensor, threshold: float = 0.5, multilabel: bool = False
+    num_classes: int,
+    preds: torch.Tensor,
+    target: torch.Tensor,
+    threshold: float = 0.5,
+    multilabel: bool = False
 ) -> Tuple[torch.Tensor, torch.Tensor]:
     """Convert preds and target tensors into one hot spare label tensors
 
@@ -262,5 +266,23 @@ def class_reduce(
         return fraction
 
     raise ValueError(
-        f"Reduction parameter {class_reduction} unknown." f" Choose between one of these: {valid_reduction}"
+        f"Reduction parameter {class_reduction} unknown."
+        f" Choose between one of these: {valid_reduction}"
     )
+
+
+def _stable_1d_sort(x: torch, N: int = 2049):
+    """
+    Stable sort of 1d tensors. Pytorch defaults to a stable sorting algorithm
+    if number of elements are larger than 2048. This function pads the tensors,
+    makes the sort and returns the sorted array (with the padding removed)
+    See this discussion: https://discuss.pytorch.org/t/is-torch-sort-stable/20714
+    """
+    if x.ndim > 1:
+        raise ValueError('Stable sort only works on 1d tensors')
+    n = x.numel()
+    if N - n > 0:
+        x_max = x.max()
+        x_pad = torch.cat([x, (x_max + 1) * torch.ones(2049 - n, dtype=x.dtype, device=x.device)], 0)
+    x_sort = x_pad.sort()
+    return x_sort.values[:n], x_sort.indices[:n]