+++ pytorch_lightning/trainer/connectors/precision_connector.py
@@ -13,8 +13,8 @@
 # limitations under the License.
 
 from pytorch_lightning import _logger as log
-from pytorch_lightning.plugins.apex import ApexPlugin
-from pytorch_lightning.plugins.native_amp import NativeAMPPlugin
+from pytorch_lightning.plugins.legacy.apex import ApexPlugin
+from pytorch_lightning.plugins.legacy.native_amp import NativeAMPPlugin
 from pytorch_lightning.utilities import _APEX_AVAILABLE, _NATIVE_AMP_AVAILABLE, AMPType, rank_zero_warn
 
 
@@ -67,7 +67,6 @@ class PrecisionConnector:
                 self.trainer.amp_backend = AMPType.APEX
                 self.backend = ApexPlugin(self.trainer)
                 log.warn("LightningOptimizer doesn't support Apex")
-                self.trainer._enable_pl_optimizer = False
 
         if not self.trainer.amp_backend:
             raise ModuleNotFoundError(